move response key
prerequisite for refactoring errors/path for connector errors
