+++ apollo-federation/src/connectors/runtime.rs
@@ -2,4 +2,5 @@ pub mod debug;
 pub mod form_encoding;
 pub mod http_json_transport;
 pub mod inputs;
+pub mod key;
 pub mod mapping;
