+++ apollo-router/src/plugins/connectors/handle_responses.rs
@@ -10,6 +10,7 @@ use apollo_federation::connectors::runtime::debug::ConnectorDebugHttpRequest;
 use apollo_federation::connectors::runtime::debug::SelectionData;
 use apollo_federation::connectors::runtime::http_json_transport::HttpResponse;
 use apollo_federation::connectors::runtime::http_json_transport::TransportResponse;
+use apollo_federation::connectors::runtime::key::ResponseKey;
 use apollo_federation::connectors::runtime::mapping::Problem;
 use apollo_federation::connectors::runtime::mapping::aggregate_apply_to_errors;
 use axum::body::HttpBody;
@@ -28,7 +29,6 @@ use tracing::Span;
 use crate::Context;
 use crate::graphql;
 use crate::json_ext::Path;
-use crate::plugins::connectors::make_requests::ResponseKey;
 use crate::plugins::telemetry::config_new::attributes::HTTP_RESPONSE_BODY;
 use crate::plugins::telemetry::config_new::attributes::HTTP_RESPONSE_HEADERS;
 use crate::plugins::telemetry::config_new::attributes::HTTP_RESPONSE_STATUS;
@@ -209,7 +209,7 @@ impl RawResponse {
                 // Now we can create the error object using either the default message or the message calculated by the JSONSelection
                 let mut error = graphql::Error::builder()
                     .message(message)
-                    .path::<Path>((&key).into());
+                    .path(Path::from_response_key(&key));
 
                 // First, we will apply defaults... these may get overwritten below by user configured extensions
                 error = error
@@ -492,7 +492,10 @@ pub(crate) async fn process_response<T: HttpBody>(
         // This occurs when we short-circuit the request when over the limit
         Err(error) => {
             let raw = RawResponse::Error {
-                error: error.to_graphql_error(connector.clone(), Some((&response_key).into())),
+                error: error.to_graphql_error(
+                    connector.clone(),
+                    Some(Path::from_response_key(&response_key)),
+                ),
                 key: response_key,
             };
             Span::current().record(OTEL_STATUS_CODE, OTEL_STATUS_CODE_ERROR);
@@ -642,7 +645,7 @@ async fn deserialize_response<T: HttpBody>(
             .with_subgraph_name(&connector.id.subgraph_name) // for include_subgraph_errors
     };
 
-    let path: Path = response_key.into();
+    let path = Path::from_response_key(response_key);
     let body = &router::body::into_bytes(body)
         .await
         .map_err(|_| make_err(path.clone()))?;
@@ -814,6 +817,7 @@ mod tests {
     use apollo_federation::connectors::HttpJsonTransport;
     use apollo_federation::connectors::JSONSelection;
     use apollo_federation::connectors::runtime::inputs::RequestInputs;
+    use apollo_federation::connectors::runtime::key::ResponseKey;
     use http::Uri;
     use insta::assert_debug_snapshot;
     use itertools::Itertools;
@@ -821,7 +825,6 @@ mod tests {
     use crate::Context;
     use crate::graphql;
     use crate::plugins::connectors::handle_responses::process_response;
-    use crate::plugins::connectors::make_requests::ResponseKey;
     use crate::services::router;
     use crate::services::router::body::RouterBody;
