+++ apollo-router/src/plugins/connectors/make_requests.rs
@@ -1,16 +1,13 @@
 use std::sync::Arc;
 
-use apollo_compiler::Name;
-use apollo_compiler::executable::FieldSet;
 use apollo_compiler::executable::Selection;
-use apollo_compiler::validation::Valid;
 use apollo_federation::connectors::Connector;
 use apollo_federation::connectors::EntityResolver;
-use apollo_federation::connectors::JSONSelection;
 use apollo_federation::connectors::runtime::debug::ConnectorContext;
 use apollo_federation::connectors::runtime::http_json_transport::HttpJsonTransportError;
 use apollo_federation::connectors::runtime::http_json_transport::make_request;
 use apollo_federation::connectors::runtime::inputs::RequestInputs;
+use apollo_federation::connectors::runtime::key::ResponseKey;
 use parking_lot::Mutex;
 
 use crate::Context;
@@ -23,55 +20,6 @@ const REPRESENTATIONS_VAR: &str = "representations";
 const ENTITIES: &str = "_entities";
 const TYPENAME: &str = "__typename";
 
-#[derive(Clone)]
-pub(crate) enum ResponseKey {
-    RootField {
-        name: String,
-        selection: Arc<JSONSelection>,
-        inputs: RequestInputs,
-    },
-    Entity {
-        index: usize,
-        selection: Arc<JSONSelection>,
-        inputs: RequestInputs,
-    },
-    EntityField {
-        index: usize,
-        field_name: String,
-        /// Is Some only if the output type is a concrete object type. If it's
-        /// an interface, it's treated as an interface object and we can't emit
-        /// a __typename in the response.
-        typename: Option<Name>,
-        selection: Arc<JSONSelection>,
-        inputs: RequestInputs,
-    },
-    BatchEntity {
-        selection: Arc<JSONSelection>,
-        keys: Valid<FieldSet>,
-        inputs: RequestInputs,
-    },
-}
-
-impl ResponseKey {
-    pub(crate) fn selection(&self) -> &JSONSelection {
-        match self {
-            ResponseKey::RootField { selection, .. } => selection,
-            ResponseKey::Entity { selection, .. } => selection,
-            ResponseKey::EntityField { selection, .. } => selection,
-            ResponseKey::BatchEntity { selection, .. } => selection,
-        }
-    }
-
-    pub(crate) fn inputs(&self) -> &RequestInputs {
-        match self {
-            ResponseKey::RootField { inputs, .. } => inputs,
-            ResponseKey::Entity { inputs, .. } => inputs,
-            ResponseKey::EntityField { inputs, .. } => inputs,
-            ResponseKey::BatchEntity { inputs, .. } => inputs,
-        }
-    }
-}
-
 /// Convert a ResponseKey into a Path for use in GraphQL errors. This mimics
 /// the behavior of a GraphQL subgraph, including the `_entities` field. When
 /// the path gets to [`FetchNode::response_at_path`], it will be amended and
@@ -80,8 +28,8 @@ impl ResponseKey {
 /// - parent path: `["posts", @, "user"]
 /// - path from key: `["_entities", 0, "user", "profile"]`
 /// - result: `["posts", 1, "user", "profile"]`
-impl From<&ResponseKey> for Path {
-    fn from(key: &ResponseKey) -> Self {
+impl Path {
+    pub(crate) fn from_response_key(key: &ResponseKey) -> Self {
         match key {
             ResponseKey::RootField { name, .. } => {
                 Path::from_iter(vec![PathElement::Key(name.to_string(), None)])
@@ -100,57 +48,6 @@ impl From<&ResponseKey> for Path {
             ResponseKey::BatchEntity { .. } => {
                 Path::from_iter(vec![PathElement::Key("_entities".to_string(), None)])
             }
-        }
-    }
-}
-
-impl std::fmt::Debug for ResponseKey {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::RootField {
-                name,
-                selection,
-                inputs,
-            } => f
-                .debug_struct("RootField")
-                .field("name", name)
-                .field("selection", &selection.to_string())
-                .field("inputs", inputs)
-                .finish(),
-            Self::Entity {
-                index,
-                selection,
-                inputs,
-            } => f
-                .debug_struct("Entity")
-                .field("index", index)
-                .field("selection", &selection.to_string())
-                .field("inputs", inputs)
-                .finish(),
-            Self::EntityField {
-                index,
-                field_name,
-                typename,
-                selection,
-                inputs,
-            } => f
-                .debug_struct("EntityField")
-                .field("index", index)
-                .field("field_name", field_name)
-                .field("typename", typename)
-                .field("selection", &selection.to_string())
-                .field("inputs", inputs)
-                .finish(),
-            Self::BatchEntity {
-                selection,
-                keys,
-                inputs,
-            } => f
-                .debug_struct("BatchEntity")
-                .field("selection", &selection.to_string())
-                .field("key_selection", &keys.serialize().no_indent().to_string())
-                .field("inputs", inputs)
-                .finish(),
         }
     }
 }
